+++ server/src/test/java/com/arcadedb/server/BaseGraphServerTest.java
@@ -391,7 +391,8 @@ public abstract class BaseGraphServerTest {
   protected void deleteDatabaseFolders() {
     if (databases != null)
       for (int i = 0; i < databases.length; ++i) {
-        databases[i].drop();
+        if (databases[i] != null)
+          databases[i].drop();
       }
 
     if (servers != null)
@@ -408,7 +409,8 @@ public abstract class BaseGraphServerTest {
   protected void deleteAllDatabases() {
     if (databases != null)
       for (int i = 0; i < databases.length; ++i) {
-        databases[i].drop();
+        if (databases[i] != null)
+          databases[i].drop();
       }
 
     if (servers != null)
