+++ server/src/test/java/performance/BasePerformanceTest.java
@@ -177,7 +177,8 @@ public abstract class BasePerformanceTest {
   protected void deleteDatabaseFolders() {
     if (databases != null)
       for (int i = 0; i < databases.length; ++i) {
-        databases[i].drop();
+        if (databases[i] != null)
+          databases[i].drop();
       }
 
     if (servers != null)
